fix: enable flexible t-of-n cryptogrpahic operations
Originally, the t parties in a t-of-n cryptographic operations
(e.g. DKG) have to be first t parties, since they were identified
by their numerical indices. In this change. We refactor job*_t to
use (unique) party name and use it as the only way to identify a
party robustly. The indices are only used within a session of a
protocol execution but not used as part of a persistent storage.

Go bindings are changed accordingly.